style: clean up client docstrings, split transaction_data module
- Move BuildTransactionData impls to router_client/transaction_data.rs
- Update ExchangeClient and StabilityPoolClient docstrings to reflect
  admin-only role
- Remove stale doc examples referencing deleted types
